requirements.txt
Django>=4.2
djangorestframework>=3.14.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.25.0